% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory_structure.R
\name{create_directory_structure}
\alias{create_directory_structure}
\title{Create Project Directories}
\usage{
create_directory_structure()
}
\value{
Create directories in the project working directory or root dirctory containin
}
\description{
Create Project Directories
}
\examples{
create_directory_structure()
}
